requirements.txt
langchain==0.0.58
tiktoken
openai
faiss-cpu
streamlit
streamlit-chat
